Started Constructing The Game Proper
Outta the debug, into the... less debug.

+++ objects/ob_text/Draw_0.gml
@@ -1,40 +1,44 @@
 // checks if mouse enter is true, then draws the amount of recipes unlocked out of max amount of recipes
 
-if (book.show_text){
-	var text = string(book.unlock_number) + "/" + string(array_length(combo));
-	var width = book.sprite_width/2;
-	var height = book.sprite_height/2;
-	smallCenteredText();
-	draw_text(book.x + width, book.y + height, text);
-}
-
-if customer.food_wanted != noone{
-	var text = "I want " + customer.food_list[customer.food_pos][0][1];
-	var width = customer.sprite_width/2;
-	smallCenteredText();
-	draw_text(customer.x + width, customer.y - offset, text);
-}
-
-// i want to keep text a temporary variable. but this keeps updating every frame
-// so i have to use a non-temporary variable
-// hence chosen
-if customer.mood != 0{
-	var text = text;
-	if chosen = noone{
-		switch customer.mood{
-			case 1:
-				text = choose("I have an army of lawyers.", "I'm literally going to kill myself.", "What's wrong with you?");
-			break;
-			case 2:
-				text = choose("Yay, thanks!", "This is the best thing ever.", "Hell yes...");
-			break;
-		}
-		chosen = text;
+if instance_exists(book){
+	if (book.show_text){
+		var text = string(book.unlock_number) + "/" + string(array_length(combo)-6);
+		var width = book.sprite_width/2;
+		var height = book.sprite_height/2;
+		smallCenteredText();
+		draw_text(book.x + width, book.y + height, text);
 	}
-	var width = string_width(chosen);
-	var width = customer.sprite_width/2;
-	smallCenteredText();
-	draw_text(customer.x + width, customer.y - offset, chosen);
+}
+
+if instance_exists(customer){
+	if customer.food_wanted != noone{
+		var text = "I want " + customer.food_list[customer.food_pos][0][1];
+		var width = customer.sprite_width/2;
+		smallCenteredText();
+		draw_text(customer.x + width, customer.y - offset, text);
+	}
+
+	// i want to keep text a temporary variable. but this keeps updating every frame
+	// so i have to use a non-temporary variable
+	// hence chosen
+	if customer.mood != 0{
+		var text = text;
+		if chosen = noone{
+			switch customer.mood{
+				case 1:
+					text = choose("I have an army of lawyers.", "I'm literally going to kill myself.", "What's wrong with you?");
+				break;
+				case 2:
+					text = choose("Yay, thanks!", "This is the best thing ever.", "Hell yes...");
+				break;
+			}
+			chosen = text;
+		}
+		var width = string_width(chosen);
+		var width = customer.sprite_width/2;
+		smallCenteredText();
+		draw_text(customer.x + width, customer.y - offset, chosen);
+	}
 }
 
 // caption underneath ingredient
@@ -49,7 +53,7 @@ if instance_exists(select){
 				width = ingredient_list[i].sprite_width/2;
 				smallCenteredText();
 				// apparently game maker can read instance refs. good to know
-				draw_text(ingredient_list[i].x + width, ingredient_list[i].y + ingredient_list[i].sprite_height, text);
+				draw_text(ingredient_list[i].x + width, ingredient_list[i].y + ingredient_list[i].sprite_height + offset/4, text);
 			}
 	}
 }
